requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
